+++ wallet/api/admin_list_networks_test.go
@@ -8,6 +8,7 @@ import (
 	"code.vegaprotocol.io/vega/libs/jsonrpc"
 	"code.vegaprotocol.io/vega/wallet/api"
 	"code.vegaprotocol.io/vega/wallet/api/mocks"
+	"code.vegaprotocol.io/vega/wallet/network"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -20,12 +21,51 @@ func TestAdminListNetworks(t *testing.T) {
 
 func testListingNetworksSucceeds(t *testing.T) {
 	// given
-	expectedNetworks := []string{"fairground", "mainnnet", "local"}
+	fairground := &network.Network{
+		Name: "fairground",
+		Metadata: []network.Metadata{
+			{
+				Key:   "category",
+				Value: "test",
+			},
+		},
+	}
+	mainnet := &network.Network{
+		Name: "mainnet",
+		Metadata: []network.Metadata{
+			{
+				Key:   "category",
+				Value: "main",
+			},
+		},
+	}
+	local := &network.Network{
+		Name: "local",
+	}
+	expectedNetworks := []api.AdminListNetworkResult{
+		{
+			Name:     fairground.Name,
+			Metadata: fairground.Metadata,
+		},
+		{
+			Name:     mainnet.Name,
+			Metadata: mainnet.Metadata,
+		},
+		{
+			Name:     local.Name,
+			Metadata: local.Metadata,
+		},
+	}
 
 	// setup
 	handler := newListNetworksHandler(t)
 	// -- expected calls
-	handler.networkStore.EXPECT().ListNetworks().Times(1).Return(expectedNetworks, nil)
+	handler.networkStore.EXPECT().ListNetworks().Times(1).Return([]string{"fairground", "mainnet", "local"}, nil)
+	gomock.InOrder(
+		handler.networkStore.EXPECT().GetNetwork("fairground").Times(1).Return(fairground, nil),
+		handler.networkStore.EXPECT().GetNetwork("mainnet").Times(1).Return(mainnet, nil),
+		handler.networkStore.EXPECT().GetNetwork("local").Times(1).Return(local, nil),
+	)
 
 	// when
 	result, errorDetails := handler.handle(t, context.Background(), nil)